X86PagingMethodPAE::MemoryTypeToPageTableEntryFlags(): Return value must be uint64, since that's the width of the page table entry. Was harmless, though, since the flags are in the lower 32 bits anyway.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@37129 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/arch/x86/paging/pae/X86PagingMethodPAE.h
@@ -91,7 +91,7 @@ public:
 									pae_page_directory_entry* const* pdpt,
 									addr_t address);
 
-	static	uint32				MemoryTypeToPageTableEntryFlags(
+	static	uint64				MemoryTypeToPageTableEntryFlags(
 									uint32 memoryType);
 
 private:
@@ -193,7 +193,7 @@ X86PagingMethodPAE::ClearPageTableEntryFlags(pae_page_table_entry* entry,
 }
 
 
-/*static*/ inline uint32
+/*static*/ inline uint64
 X86PagingMethodPAE::MemoryTypeToPageTableEntryFlags(uint32 memoryType)
 {
 	// ATM we only handle the uncacheable and write-through type explicitly. For
